fix: Prevent concurrent Google auth initialization
The `/session/authorize` may be called multiple times concurrently from
the frontend (because of a new rendering of the app triggered).
In that case, a new CSRF may be created while the frontend redirect the
user to a Google URL based on a different CSRF token. When completing
the OIDC flow, the CSRF token passed to Google was different from the
new CSRF token stored in the session (by the second call to
`/session/authorize`) and reported an error.

To fix this, when calling the `/session/authorize` service for the first
time, the authorization URL built (with a CSRF token) is also stored in
the session.
If the service is called concurrently (ie. before the previous flow is
completed), the authorization URL is found in the user's session and
directly returned without generating a new one (and a new CSRF token).
As soon as Google redirects the user to Universal Inbox to complete the
OIDC flow, the authorization URL is removed from the session.