fix: Disable auth middleware tracing
We want the auth middleware to handle the request before the tracing
middleware so that the user ID is available in the latter.
But as the tracing middleware is creating the root span, using tracing
before (ie. in the auth middleware) breaks tracing (ie.spans are not
reliably connected to each others).